feat: refactor PITR utilities and types; consolidate logic and improve structure
